vnc-proxy: Bugfix restart procedure
The proxy did not restart properly after the packer finished
and a new port was presented the old one was still active.
This change fixes the port detection and shutdown.

Basically instead of rerunning the loop start a new instance
of the script with exec. which solved the problem.

Also add help and debug options and a short description in
the usage output.